requirements.txt
openai
pandas
numpy
scikit-learn==1.2.2
langchain==0.0.148
boto3
matplotlib
plotly
scipy
google-api-python-client
pycld3